feat(auth): add user logout logic
