+++ src/app/(auth)/forgot/page.jsx
@@ -1,0 +1,5 @@
+const ForgotPassword = () => {
+  return <div>Forgot password</div>;
+};
+
+export default ForgotPassword;
