+++ src/app/feedback/page.js
@@ -1,38 +1,17 @@
-"use client";
-import React, { useState } from "react";
+import React from "react";
 import Image from "next/image";
 import styles from "./feedback.module.css";
-import { postFeedback } from "@/services";
+import FeedbackForm from "@/components/feedbackForm/feedbackForm";
 const basePath = process.env.basePath;
+import { redirect } from "next/navigation";
+import { getSession } from "@/lib/action";
 
-const Feedback = () => {
-  const [feedback, setFeedback] = useState("");
-  const [isSending, setIsSending] = useState(false);
-  const [isDisabled, setIsDisabled] = useState(true);
-  const [response, setResponse] = useState("");
+const Feedback = async () => {
+  const session = await getSession();
 
-  const btnTitle = isSending ? "Sending..." : "Send";
-
-  const handleInputChange = (event) => {
-    setFeedback(event.target.value);
-    setIsDisabled(event.target.value.length == 0);
-  };
-
-  const handleSubmit = async (event) => {
-    event.preventDefault(); // Prevents the default form submit action (page reload)
-    setIsSending(true);
-    try {
-      const result = await postFeedback(feedback);
-      setResponse(result);
-    } catch (error) {
-      console.error(error);
-      setResponse("An error occurred. Please try again later.");
-
-      // Handle error (e.g., show an error message)
-    }
-    setIsSending(false);
-  };
-
+  if (!session.isLoggedIn) {
+    redirect("/");
+  }
   return (
     <div className={styles.container}>
       <div className={styles.imgContainer}>
@@ -43,24 +22,7 @@ const Feedback = () => {
           className={styles.img}
         />
       </div>
-      <div className={styles.formContainer}>
-        <div className={styles.titleContainer}>How was your food?</div>
-        <form className={styles.form} onSubmit={handleSubmit}>
-          <textarea
-            name="comment"
-            id="comment"
-            cols="30"
-            rows="3"
-            placeholder="Tell us about your experience..."
-            value={feedback}
-            onChange={handleInputChange}
-          ></textarea>
-          {!isDisabled && <button type="submit">{btnTitle}</button>}
-        </form>
-        {!!response && (
-          <div className={styles.responseContainer}>{response}</div>
-        )}
-      </div>
+      <FeedbackForm />
     </div>
   );
 };
