+++ src/components/logoutForm/logoutForm.jsx
@@ -1,7 +1,9 @@
+import { handleLogout } from "@/lib/action";
 import styles from "./logoutForm.module.css";
+
 const LogoutForm = () => {
   return (
-    <form>
+    <form action={handleLogout}>
       <button className={styles.logout}>Logout</button>
     </form>
   );
