+++ src/components/navbar/Navbar.jsx
@@ -4,6 +4,7 @@ import styles from "./navbar.module.css";
 
 const Navbar = async () => {
   const session = await getSession();
+
   return (
     <div className={styles.container}>
       <div className={styles.logo}>REST-aurant</div>
