+++ src/components/registerForm/registerForm.module.css
@@ -22,3 +22,8 @@
   border: none;
   border-radius: 5px;
 }
+
+.error {
+  color: coral;
+  font-size: 18px;
+}
