+++ src/lib/action.js
@@ -1,11 +1,10 @@
 "use server";
 
 import { revalidatePath } from "next/cache";
-import { postRegister, postLogin } from "@/services";
+import { postRegister, getUser } from "@/services";
 import bcrypt from "bcryptjs";
 import { getIronSession } from "iron-session";
 import { cookies } from "next/headers";
-import { redirect } from "next/navigation";
 
 const sessionOptions = {
   password: process.env.SECRET_KEY,
@@ -24,14 +23,23 @@ export const getSession = async () => {
 
 export const login = async (prevState, formData) => {
   try {
-    const session = await getIronSession(cookies(), sessionOptions);
     const { email, password } = Object.fromEntries(formData);
+
+    if (!email) return { error: "Email is required!" };
+    if (!password) return { error: "Password is required!" };
+
+    const session = await getIronSession(cookies(), sessionOptions);
     console.log("Logging in user: ", email, password);
-    const user = await postLogin(email, password);
+    const user = await getUser(email);
+
+    const isPasswordCorrect = await bcrypt.compare(password, user.Password);
+
+    if (!isPasswordCorrect) return { error: "Incorrect password" };
     console.log(user);
 
     // set session properties
     session.userId = user.id;
+    session.name = user.name;
     session.isLoggedIn = true;
     await session.save();
   } catch (err) {
@@ -46,11 +54,13 @@ export const registerUser = async (previousState, formData) => {
 
   if (!name) return { error: "Name is required!" };
   if (!email) return { error: "Email is required!" };
+  if (!password) return { error: "Password is required!" };
   if (password !== passwordRepeat) return { error: "Passwords do not match!" };
 
   try {
-    console.log("Registering user: ", name, email, password);
-    await postRegister(name, email, password, "User");
+    const hashedPassword = await encryptPassword(password);
+    console.log("Registering user: ", name, email, hashedPassword);
+    await postRegister(name, email, hashedPassword, "User");
     console.log("User registered successfully...");
     return { success: true };
   } catch (err) {
@@ -61,7 +71,8 @@ export const registerUser = async (previousState, formData) => {
 
 export const handleLogout = async () => {
   //TODO
-  console.log("logout");
+  const session = await getIronSession(cookies(), sessionOptions);
+  session.destroy();
 };
 
 async function encryptPassword(password) {
